+++ CHANGELOG.md
@@ -1,5 +1,15 @@
 # Changelog
 
+## 20231216-1-002(2023-12-16)
+
+* : add iqiyi video new checkin
+
+爱奇艺签到配置说明：
+1. iqy_ck: 爱奇艺cookie 可整段 也可只保留P00001=xxx; (必须)
+2. iqiyi_dfp: 爱奇艺cookie中__dfp字段@符号前面的部分
+3. get_iqiyi_dfp: 是否请求我的api来获取参数 再去请求爱奇艺的api来获取dfp dfp类似于设备号 cookie字段内有 (选填, 不等待0或者等待1, 默认为0)
+4. sleep_await: 因观影时长同步有延迟 故建议完成任务后等待几分钟再查询 (选填, 不等待0或者等待1, 默认为1)
+
 ## 20231216-1-001(2023-12-16)
 
 * : add tx video app checkin